requirements.txt
numpy
pandas
scipy
spacy
networkx
matplotlib
transformers
pdfplumber
neo4j
python-dotenv
openai
neo4j-graphrag[openai]
langchain-text-splitters
sentence-transformers>=2.2.0
pillow>=10.0.0
pytesseract>=0.3.10
reportlab>=4.0.0
tiktoken>=0.5.0
pdf2image>=1.16.0
